# api/openapi.yaml
openapi: 3.1.0
info:
  title: Penlight Quiz v2 API
  version: 1.0.0
  description: |
    坂道ペンライトクイズ完全刷新 (Penlight Quiz v2) の公式 API 仕様書。
    全エンティティでサロゲートキー (TypeID: grp_, col_, mem_, quiz_, usr_, ans_) を採用し、
    Local-First / PWA オフライン動作および会場でのリアルタイム同期に対応します。
servers:
  - url: https://penlight.aooba.net
    description: Production Environment (Cloudflare Ingress + Homelab k8s)
  - url: http://localhost:8080
    description: Local Development Server
paths:
  /api/v1/sync/bootstrap:
    get:
      summary: オフライン用マスタ一括取得 (Bootstrap)
      description: |
        PWA クライアント起動時に、オフラインクイズ実行に必要な全グループ・カラー・メンバー情報を一括取得します。
        ETag / If-None-Match に対応し、更新がない場合は 304 Not Modified を返します。
      responses:
        "200":
          description: 最新マスタデータ
          headers:
            ETag:
              schema:
                type: string
              description: マスタデータのハッシュ値
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapData"
        "304":
          description: 更新なし (クライアントキャッシュ有効)
...
  /api/v1/quiz/answers/batch:
    post:
      summary: オフライン回答ログ一括同期
      description: 圏外・オフライン時にクライアント IndexedDB にキューイングされた回答を一括登録します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchAnswerRequest"
      responses:
        "200":
          description: 同期結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchAnswerResponse"
...
  /api/v1/quiz/statistics:
    get:
      summary: クイズ回答統計および苦手分析データ取得
      description: 回答ログから全体・グループ別・苦手対象別の正答率・平均回答時間を集計して返却します。
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ID"
        - name: group_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ID"
        - name: target_type
          in: query
          required: false
          schema:
            type: string
            enum: [member, song]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: 回答統計データ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizStatisticsResponse"
...
components:
  schemas:
    ID:
      type: string
      pattern: "^(grp|col|mem|quiz|usr|ans)_[0-9a-f]{32}$"
      description: RFC 9562 UUID v7 にエンティティ識別プレフィックスを付与したサロゲートキー (TypeID)
      example: mem_019245a28c9d7a1e8f2b3c4d5e6f7a8b
    MemberStatus:
      type: string
      enum: [active, graduated, hiatus]
      description: メンバー活動ステータス (active:現役, graduated:卒業, hiatus:休業)
      example: active
    PenlightPair:
      type: object
      required: [left_color_id, right_color_id, ordered]
      properties:
        left_color_id:
          $ref: "#/components/schemas/ID"
        right_color_id:
          $ref: "#/components/schemas/ID"
        ordered:
          type: boolean
          description: 左右の持ち順序が固定されているか (true:左右区別あり, false:順不同)
          example: false
    Group:
      type: object
      required: [id, name, short_name, slug, theme_color_hex, display_order, is_active, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        name:
          type: string
          example: 日向坂46
        short_name:
          type: string
          example: 日向坂
        slug:
          type: string
          example: hinatazaka46
        theme_color_hex:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#7CC7E8"
        display_order:
          type: integer
          example: 1
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Color:
      type: object
      required: [id, name, hex_code, display_order, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        group_id:
          $ref: "#/components/schemas/ID"
        name:
          type: string
          example: スカイブルー
        hex_code:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#00BFFF"
        display_order:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Member:
      type: object
      required: [id, group_id, family_name, given_name, family_name_kana, given_name_kana, generation, status, penlight, image_key, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        group_id:
          $ref: "#/components/schemas/ID"
        family_name:
          type: string
          example: 加藤
        given_name:
          type: string
          example: 史帆
        family_name_kana:
          type: string
          example: かとう
        given_name_kana:
          type: string
          example: しほ
        generation:
          type: integer
          example: 1
        status:
          $ref: "#/components/schemas/MemberStatus"
        penlight:
          $ref: "#/components/schemas/PenlightPair"
        image_key:
          type: string
          example: mem_019245a28c9d7a1e8f2b3c4d5e6f7a8b.webp
        joined_at:
          type: string
          format: date-time
        graduated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BootstrapData:
      type: object
      required: [groups, colors, members, generated_at]
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        colors:
          type: array
          items:
            $ref: "#/components/schemas/Color"
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        generated_at:
          type: string
          format: date-time
...
    BatchAnswerItem:
      type: object
      required: [id, quiz_question_id, group_id, is_correct, response_time_ms, answered_at]
      properties:
        id:
          $ref: "#/components/schemas/ID"
          description: クライアントで先行発行された ans_<uuidv7> (冪等性キー)
        quiz_question_id:
          $ref: "#/components/schemas/ID"
        target_member_id:
          $ref: "#/components/schemas/ID"
          description: 出題対象メンバーID (mem_...)
        target_song_id:
          $ref: "#/components/schemas/ID"
          description: 出題対象楽曲ID (sng_...)
        group_id:
          $ref: "#/components/schemas/ID"
        is_correct:
          type: boolean
        response_time_ms:
          type: integer
        answered_at:
          type: string
          format: date-time
    BatchAnswerRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/BatchAnswerItem"
    BatchAnswerResponse:
      type: object
      required: [synced_count, synced_ids]
      properties:
        synced_count:
          type: integer
          example: 5
        synced_ids:
          type: array
          items:
            $ref: "#/components/schemas/ID"
...
    TargetStat:
      type: object
      required: [target_id, target_type, group_id, name, total_answers, correct_answers, accuracy_rate, average_response_time_ms]
      properties:
        target_id:
          $ref: "#/components/schemas/ID"
        target_type:
          type: string
          enum: [member, song]
        group_id:
          $ref: "#/components/schemas/ID"
        name:
          type: string
        total_answers:
          type: integer
        correct_answers:
          type: integer
        accuracy_rate:
          type: number
          format: float
        average_response_time_ms:
          type: integer
        metadata:
          type: object
          additionalProperties: true
    GroupStat:
      type: object
      required: [group_id, group_name, total_answers, correct_answers, accuracy_rate, average_response_time_ms]
      properties:
        group_id:
          $ref: "#/components/schemas/ID"
        group_name:
          type: string
        total_answers:
          type: integer
        correct_answers:
          type: integer
        accuracy_rate:
          type: number
          format: float
        average_response_time_ms:
          type: integer
        metadata:
          type: object
          additionalProperties: true
    QuizStatisticsResponse:
      type: object
      required: [total_answers, total_correct, accuracy_rate, average_response_time_ms]
      properties:
        total_answers:
          type: integer
          example: 120
        total_correct:
          type: integer
          example: 102
        accuracy_rate:
          type: number
          format: float
          example: 0.85
        average_response_time_ms:
          type: integer
          example: 1450
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupStat"
        weak_targets:
          type: array
          items:
            $ref: "#/components/schemas/TargetStat"
        extra:
          type: object
          additionalProperties: true
...